+++ modules/features/podcasts/src/main/java/au/com/shiftyjelly/pocketcasts/podcasts/view/episode/DownloadButtonState.kt
@@ -1,0 +1,9 @@
+package au.com.shiftyjelly.pocketcasts.podcasts.view.episode
+
+sealed class DownloadButtonState {
+    data class NotDownloaded(val downloadSize: String) : DownloadButtonState()
+    object Queued : DownloadButtonState()
+    data class Downloading(val progressPercent: Float) : DownloadButtonState()
+    data class Downloaded(val downloadSize: String) : DownloadButtonState()
+    object Errored : DownloadButtonState()
+}