+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/playback/PlaybackManager.kt
@@ -66,6 +66,7 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.rx2.asFlowable
 import kotlinx.coroutines.rx2.await
 import kotlinx.coroutines.rx2.awaitSingleOrNull
 import kotlinx.coroutines.rx2.rxCompletable
@@ -366,12 +367,14 @@ open class PlaybackManager @Inject constructor(
     suspend fun playNowSync(episode: Playable, forceStream: Boolean = false, playbackSource: AnalyticsSource = AnalyticsSource.UNKNOWN) {
         LogBuffer.i(LogBuffer.TAG_PLAYBACK, "Play now: ${episode.uuid} ${episode.title}")
 
-        if (episode.isArchived) {
-            episodeManager.unarchive(episode)
-        }
+        withContext(Dispatchers.IO) {
+            if (episode.isArchived) {
+                episodeManager.unarchive(episode)
+            }
 
-        if (episode.playingStatus == EpisodePlayingStatus.COMPLETED) {
-            episodeManager.markAsNotPlayed(episode)
+            if (episode.playingStatus == EpisodePlayingStatus.COMPLETED) {
+                episodeManager.markAsNotPlayed(episode)
+            }
         }
 
         val switchEpisode: Boolean = !upNextQueue.isCurrentEpisode(episode)
@@ -391,6 +394,18 @@ open class PlaybackManager @Inject constructor(
         } else if (!switchEpisode && playbackStateRelay.blockingFirst().isPaused) {
             LogBuffer.i(LogBuffer.TAG_PLAYBACK, "No player switch required. Playing queue.")
             playQueue(playbackSource)
+        }
+    }
+
+    suspend fun play(
+        upNextPosition: UpNextPosition,
+        episode: Playable,
+        source: AnalyticsSource,
+        userInitiated: Boolean = true
+    ) {
+        when (upNextPosition) {
+            UpNextPosition.NEXT -> playNext(episode, source, userInitiated)
+            UpNextPosition.LAST -> playLast(episode, source, userInitiated)
         }
     }
 
@@ -1376,7 +1391,9 @@ open class PlaybackManager @Inject constructor(
                 return
             } else {
                 val episodeObservable: Flowable<Playable>? = if (episode is Episode) {
-                    episodeManager.observeByUuid(episode.uuid).cast(Playable::class.java)
+                    episodeManager.observeByUuid(episode.uuid)
+                        .asFlowable()
+                        .cast(Playable::class.java)
                 } else if (episode is UserEpisode) {
                     userEpisodeManager.observeEpisode(episode.uuid).cast(Playable::class.java)
                 } else {